requirements.txt
opencv-python
torch
torchvision
facenet-pytorch
numpy
scikit-learn
pillow
tk